# docs/swagger.yaml
basePath: /api/v1
definitions:
  github_com_gopal-chhetri_url-shortener_internal_response.PaginatedResponse:
    properties:
      data: {}
      error:
        type: boolean
      message:
        type: string
      metadata:
        $ref: '#/definitions/github_com_gopal-chhetri_url-shortener_internal_utils.PaginationResponse'
    type: object
  github_com_gopal-chhetri_url-shortener_internal_response.Response:
    properties:
      data: {}
      error:
        type: boolean
      message:
        type: string
    type: object
  github_com_gopal-chhetri_url-shortener_internal_utils.PaginationResponse:
    properties:
      page:
        type: integer
      page_size:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
    type: object
  internal_admin.DashboardStats:
    properties:
      active_urls:
        type: integer
      inactive_urls:
        type: integer
      top_urls:
        items:
          $ref: '#/definitions/internal_admin.TopURL'
        type: array
      total_clicks:
        type: integer
      total_users:
        type: integer
    type: object
  internal_admin.TopURL:
    properties:
      click_count:
        type: integer
      id:
        type: string
      original_url:
        type: string
      short_url:
        type: string
    type: object
  internal_admin.UpdateRoleRequest:
    properties:
      role:
        type: string
    required:
    - role
    type: object
  internal_admin.UpdateStatusRequest:
    properties:
      is_active:
        type: boolean
    type: object
  internal_admin.UserResponse:
    properties:
      created_at:
        type: string
      email:
        type: string
      first_name:
        type: string
      id:
        type: string
      is_active:
        type: boolean
      last_name:
        type: string
      role:
        type: string
    type: object
...
host: localhost:8000
info:
  contact: {}
  description: This is a sample URL Shortener server with authentication and authorization.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: URL Shortener API
  version: "1.0"
paths:
...
  /admin/roles:
    get:
      description: Returns all available user roles
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_gopal-chhetri_url-shortener_internal_response.Response'
      security:
      - BearerAuth: []
      summary: List all roles (admin)
      tags:
      - admin
  /admin/stats:
    get:
      description: 'Returns aggregate stats: total users, active/inactive URLs, total
        clicks, top 5 URLs'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_gopal-chhetri_url-shortener_internal_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/internal_admin.DashboardStats'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_gopal-chhetri_url-shortener_internal_response.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_gopal-chhetri_url-shortener_internal_response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_gopal-chhetri_url-shortener_internal_response.Response'
      security:
      - BearerAuth: []
      summary: Get admin dashboard statistics
      tags:
      - admin
  /admin/users:
    get:
      description: Get a paginated list of all users in the system
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Items per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_gopal-chhetri_url-shortener_internal_response.PaginatedResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/internal_admin.UserResponse'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_gopal-chhetri_url-shortener_internal_response.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_gopal-chhetri_url-shortener_internal_response.Response'
      security:
      - BearerAuth: []
      summary: List all users (admin)
      tags:
      - admin
  /admin/users/{id}/role:
    put:
      consumes:
      - application/json
      description: Assign a new role to a user
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - description: Role update request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_admin.UpdateRoleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_gopal-chhetri_url-shortener_internal_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/internal_admin.UserResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_gopal-chhetri_url-shortener_internal_response.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_gopal-chhetri_url-shortener_internal_response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_gopal-chhetri_url-shortener_internal_response.Response'
      security:
      - BearerAuth: []
      summary: Update user role (admin)
      tags:
      - admin
  /admin/users/{id}/status:
    put:
      consumes:
      - application/json
      description: Activate or deactivate a user account
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - description: Status update request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_admin.UpdateStatusRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_gopal-chhetri_url-shortener_internal_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/internal_admin.UserResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_gopal-chhetri_url-shortener_internal_response.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_gopal-chhetri_url-shortener_internal_response.Response'
      security:
      - BearerAuth: []
      summary: Toggle user active status (admin)
      tags:
      - admin
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"